added a manifest compliant load and clean function. To make this work you must have a manifest. You may wish to combine the ICOS and AMF manifests (they can be catted together). Other new functions are revised versions of the discover_AMF_files and discover_ICOS_files (these build the manifest)
There is a newer unzip function - i needed to run this a few time before I unzipped all 240 ish AMF sites.

There is also an organize_fluxnet_files() function that will stow all the FLUXNET sites in an appropriate file system. This isn't strictly needed as the load function now just reads the manifest but it made sense to me.

need to test the loading function a bit more but it seems to load the annual data fine